requirements.txt
asyncio==3.4.3
attrs==20.3.0
autopep8==2.0.0
awscli==1.27.18
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
beautifulsoup4==4.11.1
botocore==1.29.18
braceexpand==0.1.7
Brotli==1.0.9
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
check50==3.3.7
click==8.1.3
colorama==0.4.4
compare50==1.2.6
cryptography==38.0.4
cs50==9.2.2
cssselect2==0.7.0
docutils==0.16
EditorConfig==0.12.3
Flask==2.2.2
Flask-Session==0.4.0
fonttools==4.38.0
greenlet==2.0.1
help50==3.0.5
html5lib==1.1
icdiff==2.0.5
idna==3.4
iniconfig==1.1.1
intervaltree==3.1.0
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.7
lib50==3.0.4
MarkupSafe==2.1.1
natsort==8.2.0
numpy==1.23.5
packaging==21.3
pexpect==4.8.0
Pillow==9.3.0
pluggy==1.0.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.5.0
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==2.11.2
pyphen==0.13.2
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-magic==0.4.27
PyYAML==5.4.1
render50==9.0.2
requests==2.28.1
rsa==4.7.2
s3cmd==2.3.0
s3transfer==0.6.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
style50==2.7.5
submit50==3.1.1
termcolor==1.1.0
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.13
weasyprint==57.1
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
zopfli==0.2.2
